fixed #66; added confirmation dialog when deleting photo

+++ public/javascripts/zz/pages.js
@@ -150,17 +150,22 @@ pages.edit_album_tab = {
                     cellHeight: 230,
 
                     onDelete: function(index, photo){
-                        $.ajax({
-                            type: "DELETE",
-                            dataType: "json",
-                            url: "/photos/" + photo.id + ".json",
-                            error: function(error){
-                                logger.debug(error);
-//                                $.jGrowl("" + error);
-                            }
-                            
-                        });
-                        return true;                          
+                        if(confirm('Are you sure you want to delete this photo?')){
+                            $.ajax({
+                                type: "DELETE",
+                                dataType: "json",
+                                url: "/photos/" + photo.id + ".json",
+                                error: function(error){
+                                    logger.debug(error);
+    //                                $.jGrowl("" + error);
+                                }
+
+                            });
+                            return true;
+                        }
+                        else{
+                            return false;
+                        }
                     },
                     allowEditCaption: true,
                     onChangeCaption: function(index, photo, caption){
